typewriter: fix last written line backlash removal
If we remove the last line of a rule, the previous one will have the
backslash if not a single rule (not covered yet and not needed).

E.x:

PRODUCT_COPY += \
    qcdm_3232.xml:system/somewhere.xml \
    qcdm_4343.xml:system/somewhere.xml

In case we want to move only qcdm_4343 to Android.mk, the previous line
would keep the backlash which is not an option on AOSP and it's not
correct either.

Workaround that little issue by grepping the previous and next lines for
more information on the strings and remove if the next is empty (last
line).

Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>